'Code_Formatting'

+++ lib/Widgets/UserProductItem.dart
@@ -16,18 +16,8 @@ class UserProductItem extends StatelessWidget {
         title: const Text('Are you sure?'),
         content: Text('You are going to remove $title from product list.'),
         actions: [
-          ElevatedButton(
-            child: const Text('No'),
-            onPressed: () {
-              Navigator.of(ctx).pop(false);
-            },
-          ),
-          ElevatedButton(
-            child: const Text('Yes'),
-            onPressed: () {
-              Navigator.of(ctx).pop(true);
-            },
-          ),
+          ElevatedButton(child: const Text('No'), onPressed: () => Navigator.of(ctx).pop(false)),
+          ElevatedButton(child: const Text('Yes'), onPressed: () => Navigator.of(ctx).pop(true)),
         ],
       ),
     );
@@ -38,31 +28,16 @@ class UserProductItem extends StatelessWidget {
   Widget build(BuildContext context) {
     return ListTile(
       title: Text(title),
-      leading: CircleAvatar(
-        backgroundImage: NetworkImage(imageUrl),
-        backgroundColor: Theme.of(context).primaryColor,
-      ),
+      leading: CircleAvatar(backgroundImage: NetworkImage(imageUrl), backgroundColor: Theme.of(context).primaryColor),
       trailing: Container(
         width: 100,
         child: Row(
           children: [
-            IconButton(
-              icon: Icon(Icons.edit),
-              onPressed: () {
-                Navigator.pushNamed(
-                  context,
-                  EditProductScreen.routeName,
-                  arguments: id,
-                );
-              },
-              color: Theme.of(context).primaryColor,
-            ),
+            IconButton(icon: Icon(Icons.edit), onPressed: () => Navigator.pushNamed(context, EditProductScreen.routeName, arguments: id), color: Theme.of(context).primaryColor),
             IconButton(
               icon: Icon(Icons.delete),
               onPressed: () async {
-                if (await deleteConfirmation(context))
-                  Provider.of<Products>(context, listen: false)
-                      .deleteProduct(id);
+                if (await deleteConfirmation(context)) Provider.of<Products>(context, listen: false).deleteProduct(id);
               },
               color: Theme.of(context).errorColor,
             ),
